Make the edge weight size_t
It reflects the number of bytes in the file, which is already being
tracked with size_t variables anyway.

+++ clownlzss.h
@@ -32,7 +32,7 @@ typedef struct ClownLZSS_GraphEdge
 {
 	union
 	{
-		unsigned int cost;
+		size_t cost;
 		size_t next_node_index;
 	} u;
 	size_t previous_node_index;
